requirements.txt
pygithub==2.6.0
pandas==2.2.3
requests==2.32.3
matplotlib==3.10.0
python-dotenv==1.0.1
seaborn==0.13.2